refactor: 리그셀렉터 shared로 분리 및 이동
